+++ griddedio/GriddedIO.F90
@@ -303,7 +303,7 @@ module MAPL_GriddedIOMod
         else if (fieldRank==3) then
            vdims=grid_dims//",lev,time"
         else
-           _ASSERT(.false., 'Unsupported field rank')
+           _FAIL( 'Unsupported field rank')
         end if
         v = Variable(type=PFIO_REAL32,dimensions=vdims,chunksizes=this%chunking,deflation=this%deflateLevel)
         call v%add_attribute('units',trim(units))
@@ -543,7 +543,7 @@ module MAPL_GriddedIOMod
               _VERIFY(status)
            end if
         else
-           _ASSERT(.false.,'rank not supported')
+           _FAIL('rank not supported')
         end if
 
         if (allocated(ptr3d_inter)) deallocate(ptr3d_inter)
@@ -853,7 +853,7 @@ module MAPL_GriddedIOMod
          allocate(globalStart,source=[gridGlobalStart,1,tindex])
          allocate(globalCount,source=[gridGlobalCount,lm,1])
       else
-         _ASSERT(.false., "Rank not supported")
+         _FAIL( "Rank not supported")
       end if
       call oClients%collective_stage_data(this%write_collection_id,trim(filename),trim(fieldName), &
            ref,start=localStart, global_start=GlobalStart, global_count=GlobalCount)
@@ -1110,7 +1110,7 @@ module MAPL_GriddedIOMod
            end if
            where(ptr3d==fill_value) ptr3d=MAPL_UNDEF
         else
-           _ASSERT(.false.,'rank not supported')
+           _FAIL('rank not supported')
         end if
      end if
      _RETURN(_SUCCESS)